AD-1: refactor and implement ontoMany according to the JPA docs
