+++ src/main/java/com/spring/rest/webservices/restwebservices/repository/PostRepository.java
@@ -1,6 +1,7 @@
-package com.spring.rest.webservices.restwebservices.user;
+package com.spring.rest.webservices.restwebservices.repository;
 
 
+import com.spring.rest.webservices.restwebservices.model.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
