+++ src/main/java/com/spring/rest/webservices/restwebservices/repository/QueueRepository.java
@@ -1,5 +1,6 @@
-package com.spring.rest.webservices.restwebservices.member;
+package com.spring.rest.webservices.restwebservices.repository;
 
+import com.spring.rest.webservices.restwebservices.model.Queue;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
