+++ src/main/java/com/spring/rest/webservices/restwebservices/repository/UserRepository.java
@@ -1,6 +1,7 @@
-package com.spring.rest.webservices.restwebservices.user;
+package com.spring.rest.webservices.restwebservices.repository;
 
 
+import com.spring.rest.webservices.restwebservices.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
